# api_metacritic_game/openapi.yaml
openapi: 3.0.1
info:
  description: "Projet Final"
  version: "1.0.6"
  title: "MetaCritic"
...
paths:
  /users:
    get:
      tags:
        - Users
      summary: Get all users
      description: Returns all users
      operationId: "getUsers" 
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                 $ref: '#/components/schemas/Users'

    post:
      tags:
        - Users
      summary: Add user
      description: Add user
      operationId: "addUser" 
      requestBody:
        $ref: '#/components/requestBodies/User'
      responses:
        '200':
          description: users response
          content:
            application/json:
              schema:
                type: array
                items:
                 $ref: '#/components/schemas/Users'

    put:
      tags:
        - Users
      summary: Update user by id
      description: Update a user by his id
      operationId: "updateUser"
      requestBody:
        $ref: '#/components/requestBodies/User'
      responses:
        '200':
          description: users response
          content:
            application/json:
              schema:
                type: array
                items:
                 $ref: '#/components/schemas/Users'
  /users/{id}:
    delete:
      tags:
        - Users
      summary: Delete user by id
      description: Delete a user by his id
      operationId: "DeleteUserById" 
      parameters:
        - $ref: '#/components/parameters/idInPath'
      responses:
        '200':
          description: users response
          content:
            application/json:
              schema:
                type: array
                items:
                 $ref: '#/components/schemas/Users'

    get:
      tags:
        - Users
      summary: Get user by id
      description: Returns a user by id
      operationId: "getUserById" 
      parameters:
        - $ref: '#/components/parameters/idInPath'
      responses:
        '200':
          description: users response
          content:
            application/json:
              schema:
                type: array
                items:
                 $ref: '#/components/schemas/Users'

    put:
      tags:
        - Users
      summary: Update role's user by id
      description: Returns new role and id
      operationId: "updateRoleByUserId" 
      parameters:
        - $ref: '#/components/parameters/idInPath'
      responses:
        '200':
          description: users response
          content:
            application/json:
              schema:
                type: array
                items:
                 $ref: '#/components/schemas/Users'
...
  /reviews:
    get:
      tags:
        - Reviews
      summary: get all reviews
      description: Returns all the reviews mayde by editors
      operationId: "getReviews" 
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                 $ref: '#/components/schemas/Reviews'

    post:
      tags:
        - Reviews
      summary: Add review
      description: Add review
      operationId: "addReview" 
      requestBody:
        $ref: '#/components/requestBodies/Review'
      responses:
        '200':
          description: review response
          content:
            application/json:
              schema:
                type: array
                items:
                 $ref: '#/components/schemas/Reviews'

    put:
      tags:
        - Reviews
      summary: Update review by id
      description: Update a review by his id
      operationId: "updateReview" 
      requestBody:
        $ref: '#/components/requestBodies/Review'
      responses:
        '200':
          description: review response
          content:
            application/json:
              schema:
                type: array
                items:
                 $ref: '#/components/schemas/Reviews'
  /reviews/{id}:
    get:
      tags:
        - Reviews
      summary: Get review by id
      description: Returns a review by his id
      operationId: "getreviewById" 
      parameters:
        - $ref: '#/components/parameters/idInPath'
      responses:
        '200':
          description: review response
          content:
            application/json:
              schema:
                type: array
                items:
                 $ref: '#/components/schemas/Reviews'
                 
    delete:
      tags:
        - Reviews
      summary: Delete review by id
      description: Delete a review by his id
      operationId: "DeleteReviewById" 
      parameters:
        - $ref: '#/components/parameters/idInPath'
      responses:
        '200':
          description: review response
          content:
            application/json:
              schema:
                type: array
                items:
                 $ref: '#/components/schemas/Reviews'
...
components:
  parameters:
    idInPath:
      required: true
      name: id
      in: path
      schema:
        type: integer
  requestBodies:
    User:
      description: A JSON object containing an user
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Users'
          examples:
            userExample:
              $ref: '#/components/examples/Users'
...
    Review:
      description: A JSON object containing a review
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/examples/Reviews'
          examples:
            reviewsExample:
              $ref: '#/components/examples/Reviews'
...
  schemas:
    Users:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 3
        pseudo:
          type: string
          example: johnDoe
        hashedPassword:
          type: string
          example: $2b$10$t7oxiwchWGHa/B9w0AzrYO2WH2rQbA86YSuQjSTmwIrpC/0ZXN7V2
        email:
          type: string
          example: johnDoe@gmail.com
...
    Reviews:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        avg:
          type: number 
          format: int64
          example: 4.4     
        description:
          type: string
          example: bon jeu 
        title:
          type: string
          example: the hobbit
        release:
          type: string
          example: 11/02/2002
        idCriticEditor:
          type: integer
          format: int64
          example: 1
...
  examples:
    Reviews:
      summary: An example of a review
      value:
        id: 1
        avg: 15
        description: "très bon jeu"
        title: "Persona 5"
        release: "11-02-2002"
        idCriticEditor: 2
...
    Users:
      summary: An example of a user
      value:
        id: 3
        pseudo: "johnDoe"
        email: "johnDoe@gmail.com"
        hashedPassword: "$2b$10$t7oxiwchWGHa/B9w0AzrYO2WH2rQbA86YSuQjSTmwIrpC/0ZXN7V2"
        roles: "user"
...